Rename bloom module to blooming_history
Naming conflict with libbloom
