+++ blooming_history/blooming_history_functions.c
@@ -13,14 +13,14 @@
  * modification, are permitted provided that the following conditions
  * are met:
  * 1. Redistributions of source code must retain the above copyright
- *    notice, this list of conditions and the following disclaimer.
+ *   notice, this list of conditions and the following disclaimer.
  * 2. Redistributions in binary form must reproduce the above copyright
- *    notice, this list of conditions and the following disclaimer in
- *    the documentation and/or other materials provided with the
- *    distribution.
+ *   notice, this list of conditions and the following disclaimer in
+ *   the documentation and/or other materials provided with the
+ *   distribution.
  * 3. Neither the name of the Company nor the names of its contributors
- *    may be used to endorse or promote products derived from this
- *    software without specific prior written permission.
+ *   may be used to endorse or promote products derived from this
+ *   software without specific prior written permission.
  *
  * ALTERNATIVELY, provided that this notice is retained in full, this
  * product may be distributed under the terms of the GNU General Public
@@ -43,33 +43,33 @@
 
 #include <unirec/unirec.h>
 
-#include "bloom_functions.h"
+#include "blooming_history_functions.h"
 
 
 int is_from_prefix(ip_addr_t * ip, ip_addr_t * protected_prefix, int32_t protected_prefix_length) {
-    // Both IPv4
-    if(ip_is4(ip) && ip_is4(protected_prefix)) {
-        uint32_t mask = 0xffffffff << (32 - protected_prefix_length);
-        return (ip_get_v4_as_int(ip) & mask) == (ip_get_v4_as_int(protected_prefix) & mask);
-    }
-    // Both IPv6
-    if (ip_is6(ip) && ip_is6(protected_prefix)) {
-        int bytes_match;
+   // Both IPv4
+   if(ip_is4(ip) && ip_is4(protected_prefix)) {
+      uint32_t mask = 0xffffffff << (32 - protected_prefix_length);
+      return (ip_get_v4_as_int(ip) & mask) == (ip_get_v4_as_int(protected_prefix) & mask);
+   }
+   // Both IPv6
+   if (ip_is6(ip) && ip_is6(protected_prefix)) {
+      int bytes_match;
 
-        // Compare whole bytes
-        bytes_match = memcmp((const char *)ip, (const char *)protected_prefix, protected_prefix_length/8) == 0;
+      // Compare whole bytes
+      bytes_match = memcmp((const char *)ip, (const char *)protected_prefix, protected_prefix_length/8) == 0;
 
-        // Compare remaining byte
-        if (bytes_match && protected_prefix_length % 8 != 0) {
-            int byte_index = protected_prefix_length/8;
-            uint8_t mask = 0xff << (8 - (protected_prefix_length % 8));
+      // Compare remaining byte
+      if (bytes_match && protected_prefix_length % 8 != 0) {
+         int byte_index = protected_prefix_length/8;
+         uint8_t mask = 0xff << (8 - (protected_prefix_length % 8));
 
-            return bytes_match && ((ip->bytes[byte_index] & mask) == (protected_prefix->bytes[byte_index] & mask));
-        }
+         return bytes_match && ((ip->bytes[byte_index] & mask) == (protected_prefix->bytes[byte_index] & mask));
+      }
 
-        return bytes_match;
-    }
+      return bytes_match;
+   }
 
-    return 0;
+   return 0;
 }
